Refactor constants in Ceph Plugin
