+++ frontend/packages/ceph-storage-plugin/src/components/ocs-install/attached-devices/create-sc/wizard-pages/configure-step.tsx
@@ -3,7 +3,7 @@ import { Form } from '@patternfly/react-core';
 import { useFlag } from '@console/shared';
 import { State, Action } from '../state';
 import { EncryptionFormGroup, NetworkFormGroup } from '../../../install-wizard/configure';
-import { NetworkType } from '../../../types';
+import { NetworkType } from '../../../../../types';
 import { GUARDED_FEATURES } from '../../../../../features';
 
 export const Configure: React.FC<ConfigureProps> = ({ state, dispatch, mode }) => {
