+++ frontend/packages/ceph-storage-plugin/src/components/ocs-install/internal-mode/install-wizard-steps/review-and-create.tsx
@@ -18,7 +18,7 @@ import {
   NodesCard,
   RequestErrors,
 } from '../../install-wizard/review-and-create';
-import { NetworkType } from '../../types';
+import { NetworkType } from '../../../../types';
 import { GUARDED_FEATURES } from '../../../../features';
 
 export const ReviewAndCreate: React.FC<ReviewAndCreateProps> = ({
