+++ frontend/packages/ceph-storage-plugin/src/constants/disks.ts
@@ -1,0 +1,3 @@
+export const diskModeDropdownItems = Object.freeze({
+  BLOCK: 'Block',
+});
