+++ frontend/packages/ceph-storage-plugin/src/constants/pool.ts
@@ -1,0 +1,4 @@
+export const OCS_DEVICE_REPLICA = Object.freeze({
+  '2': '2-way',
+  '3': '3-way',
+});
